Moving methods which are not generic page methods from page.ts to pageHelper.ts
