+++ test/e2e/pages/pageHelper.ts
@@ -3,6 +3,8 @@ import JournalPage from './journalPage';
 import LoginPage from './loginPage';
 import LandingPage from './landingPage';
 import DashboardPage from './dashboardPage';
+import { browser, Key } from 'protractor';
+import { TEST_CONFIG } from '../test.config';
 
  /**
  * Temporary page to allow access to functions in Page base class
@@ -28,6 +30,40 @@ class PageHelper extends Page {
      // Navigate to journal page
     DashboardPage.clickGoToMyJournalButton();
     JournalPage.isCurrentPage();
+  }
+
+  getPassCodeField() {
+    const element = this.getElementByXPath('//XCUIElementTypeSecureTextField[@label="Passcode field"]');
+    this.waitForPresenceOfElement(element);
+    return element;
+  }
+
+   /**
+    * Send the fake passcode using native browser actions
+    */
+  sendFakePasscode() {
+    browser.actions().sendKeys('PASSWORD').sendKeys(Key.ENTER).perform();
+  }
+
+   /**
+    * Enters a generic password into the iOS passcode field.
+    * Note: This will not work on the physical device but the simulator will accept any code.
+    */
+  enterPasscode() {
+     // To be able to fill in the passcode we need to switch to NATIVE context then switch back to WEBVIEW after
+    browser.driver.getCurrentContext().then((webviewContext) => {
+       // Switch to NATIVE context
+      browser.driver.selectContext('NATIVE_APP').then(() => {
+         // Get the passcode field
+        this.getPassCodeField();
+        this.sendFakePasscode();
+
+         // Switch back to WEBVIEW context
+        browser.driver.selectContext(this.getParentContext(webviewContext)).then(() => {
+          browser.driver.sleep(TEST_CONFIG.PAGE_LOAD_WAIT);
+        });
+      });
+    });
   }
 }
